requirements.txt
aiofiles
PyYAML
